Alternative Rust solution for day 19
The major problem with this kind of algorithm is that if the reduction
isn't straightforward, then we potentially need to search through a much
larger space to find the correct answer. Luckily for me, the correct
answer was the first one I found any way.

If, for example, we have a reduction step that removes 10 characters
from an input of length 100, then follow that with 89 single-character
removals to end at "e" we will have taken 90 steps. However, if we have
a reduction step that removes 5 characters from that same input, then
follow that with 18 5-character reductions and 1 4-character reduction,
we will have taken 20 steps.

The algorithm used is greedy - it will search the space with shorter
strings first. It will, however, explore the entire search space given
enough time and memory.